requirements.txt
pandas
langchain
langchain-community
langchain-openai
faiss-cpu
tqdm
tiktoken
numpy
openai
python-dotenv
openpyxl